requirements.txt
icrawler==0.6.10
mediapipe==0.10.14
pandas
numpy<2
customtkinter
Pillow
scikit-learn
joblib
opencv-contrib-python<4.9
